+++ Excersise 3 - SES& DES /excersise-3/src/main/java/com/hro/cmi/Vector2.java
@@ -2,10 +2,10 @@ package com.hro.cmi;
 
 class Vector2
 {
-    public float x;
-    public float y;    
+    public double x;
+    public double y;    
 
-    public Vector2(float x, float y)
+    public Vector2(double x, double y)
     {
         this.x = x;
         this.y = y;
